# fix: make asString monomorphic  and trigger V8 optimization