--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -99,20 +99,6 @@ module.exports = class Serializer {
   }
 
   asString (str) {
-    if (typeof str !== 'string') {
-      if (str === null) {
-        return '""'
-      }
-      if (str instanceof Date) {
-        return '"' + str.toISOString() + '"'
-      }
-      if (str instanceof RegExp) {
-        str = str.source
-      } else {
-        str = str.toString()
-      }
-    }
-
     if (str.length < 42) {
       return this.asStringSmall(str)
     } else if (STR_ESCAPE.test(str) === false) {