@kibertoad 2024-03-10
@Uzlopak checked https://github.com/fastify/fast-json-stringify/blob/master/types/index.d.ts, it doesn't seem to be there. so it's an internal only method?